+++ app/front/app.py
@@ -14,12 +14,12 @@ import tempfile
 import os
 
 # Import des classes métier existantes
-from back.Automate import Automate, ADC, AFDC, AND, AFND, AFNS
-from back.Etat import Etat
-from back.Mot import Mot
-from back.Langage import Langage
-from back.LangageReconnaissable import LangageReconnaissable
-from back.Monoids import Monoids
+from Automate import Automate, ADC, AFDC, AND, AFND, AFNS
+from Etat import Etat
+from Mot import Mot
+from Langage import Langage
+from LangageReconnaissable import LangageReconnaissable
+from Monoids import Monoids
 
 
 class AutomateVisualizer(ABC):
