+++ help.md
@@ -1,0 +1,71 @@
+# Guide d'implémentation des méthodes
+
+## AutomateVisualizer
+
+- **generer_svg_automate** : Utilisez une bibliothèque comme `svgwrite` pour créer des nœuds et des transitions SVG représentant les états et les transitions de l'automate.
+- **generer_canvas_html** : Créez un élément HTML5 Canvas et utilisez JavaScript pour dessiner l'automate.
+- **generer_js_animation_reconnaissance** : Écrivez du JavaScript pour animer le parcours d'un mot dans l'automate, en mettant en évidence les états actifs.
+- **generer_js_animation_determinisation** : Utilisez JavaScript pour montrer la construction pas à pas des sous-ensembles lors de la déterminisation.
+- **generer_css_styles** : Définissez des styles CSS pour les éléments SVG ou Canvas afin d'améliorer la lisibilité et l'esthétique.
+- **calculer_positions_etats** : Implémentez un algorithme de placement automatique, comme un algorithme de force dirigée, pour positionner les états.
+
+## SVGVisualizer
+
+- **generer_svg_automate** : Générez un document SVG complet en utilisant les méthodes de création d'états et de transitions.
+- **generer_svg_etat** : Créez un groupe SVG pour chaque état, avec un cercle et un texte pour l'étiquette.
+- **generer_svg_transition** : Dessinez une ligne ou une courbe SVG entre deux états, avec une étiquette pour le symbole de transition.
+
+## CanvasVisualizer
+
+- **generer_canvas_html** : Créez un élément Canvas HTML et initialisez le contexte 2D pour le dessin.
+- **generer_js_dessin_automate** : Utilisez l'API Canvas pour dessiner les états et les transitions de l'automate.
+- **generer_js_dessin_etat** : Dessinez un cercle pour représenter un état et ajoutez une étiquette textuelle.
+- **generer_js_dessin_transition** : Dessinez une flèche entre deux états pour représenter une transition, avec une étiquette pour le symbole.
+
+## RegexParser
+
+- **parser_regex** : Utilisez une bibliothèque ou un algorithme pour convertir une expression régulière en un automate équivalent.
+- **valider_syntaxe** : Vérifiez la syntaxe de l'expression régulière en utilisant des expressions régulières ou un parseur.
+- **construire_automate_base** : Créez un automate simple avec un seul état initial et final pour un symbole donné.
+
+## HTMLGenerator
+
+- **creer_document_base** : Utilisez la bibliothèque `dominate` pour créer un document HTML de base avec les balises nécessaires.
+- **ajouter_css_bootstrap** : Ajoutez un lien vers la feuille de style Bootstrap pour un design réactif.
+- **ajouter_js_bibliotheques** : Ajoutez des scripts pour des bibliothèques JavaScript comme jQuery ou D3.js.
+
+## ZoneSaisieHTML
+
+- **generer_formulaire_regex** : Créez un formulaire HTML avec un champ de saisie pour les expressions régulières.
+- **generer_formulaire_automate_manuel** : Fournissez des champs pour saisir manuellement les états et les transitions d'un automate.
+- **generer_zone_mots_test** : Ajoutez une zone de texte pour saisir des mots à tester sur l'automate.
+
+## ZoneVisualisationHTML
+
+- **generer_conteneur_visualisation** : Créez un conteneur div pour contenir les éléments de visualisation de l'automate.
+- **generer_controles_visualisation** : Ajoutez des boutons et des curseurs pour contrôler le zoom et le déplacement dans la visualisation.
+- **generer_canvas_principal** : Créez un élément Canvas principal pour dessiner l'automate.
+
+## ZoneResultatsHTML
+
+- **generer_onglets_resultats** : Utilisez Bootstrap pour créer des onglets affichant différents types de résultats.
+- **generer_tableau_proprietes** : Créez un tableau HTML pour afficher les propriétés de l'automate.
+- **generer_tableau_tests** : Affichez les résultats des tests de mots dans un tableau HTML.
+
+## GestionnaireOperations
+
+- **regex_vers_automate** : Utilisez `RegexParser` pour convertir une expression régulière en automate.
+- **determiniser_automate** : Implémentez l'algorithme de déterminisation pour convertir un automate non déterministe en un automate déterministe.
+- **minimiser_automate** : Utilisez l'algorithme de minimisation pour réduire le nombre d'états dans un automate déterministe.
+
+## ServeurLocal
+
+- **demarrer_serveur** : Utilisez une bibliothèque comme `Flask` ou `FastAPI` pour démarrer un serveur HTTP local.
+- **generer_api_endpoints** : Définissez des routes API pour convertir des regex, transformer des automates, et tester des mots.
+- **ouvrir_navigateur** : Utilisez le module `webbrowser` pour ouvrir automatiquement l'interface dans le navigateur par défaut.
+
+## Application
+
+- **initialiser_serveur** : Configurez et démarrez le serveur web en utilisant `ServeurLocal`.
+- **generer_application_complete** : Combinez tous les composants pour générer l'application web complète.
+- **lancer_application** : Démarrez le serveur et ouvrez l'interface dans le navigateur.
